Lots of Stuff
Got rid of the annoying warnings about the animations. Implemented the
removal of asteroids when they touch the Sun/go off the screen. Started
on swiping asteroids, haven’t tested it yet because I need to build it
on my phone first. Also removed a couple of images we won’t be using.